tests/wpt/tests/html/semantics/embedded-content/the-canvas-element/size.attributes.reflect.setidl.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: size.attributes.reflect.setidl</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>size.attributes.reflect.setidl</h1>
<p class="desc">Setting IDL attributes updates IDL and content attributes</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.reflect.setidl.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Setting IDL attributes updates IDL and content attributes");
_addTest(function(canvas, ctx) {

canvas.width = 120;
canvas.height = 60;
_assertSame(canvas.getAttribute('width'), '120', "canvas.getAttribute('width')", "'120'");
_assertSame(canvas.getAttribute('height'), '60', "canvas.getAttribute('height')", "'60'");
_assertSame(canvas.width, 120, "canvas.width", "120");
_assertSame(canvas.height, 60, "canvas.height", "60");


});
</script>
